python_piscine_01/requirements.txt
pillow==10.3.0
pyautogui==0.9.54
